Added currentBlock and dir enum types

+++ src/electronics.cpp
@@ -5,10 +5,10 @@
 #include "state.h"
 
 // Variables for sensors
-#define ls1 13     //L
-#define ls2 12     //Middle sensor for single line following, LM for 4 sensor following
-#define ls3 10     //RM
-#define ls4 11     //R
+#define ls1 13    //L
+#define ls2 12    //Middle sensor for single line following, LM for 4 sensor following
+#define ls3 10    //RM
+#define ls4 11    //R
 #define srange A0 // Long range sensor
 #define lrange A1 // Short range sensor
 
@@ -104,7 +104,7 @@ void pauseButton()
     }
 }
 
-int lrangeDistance()
+float lrangeDistance()
 {
     // long range distance reading
     // returns an integer giving the distance reading in mm
@@ -116,12 +116,11 @@ int lrangeDistance()
     return analogRead(lrange);
 }
 
-int srangeDistance()
+float srangeDistance()
 {
     // short range distance reading
     // returns an integer giving the distance reading in mm
-   float distance;
-   distance=-4.9261+33.4525/(float(analogRead(srange)*5)/1024);
-
+    float distance;
+    distance = -4.9261 + 33.4525 / (float(analogRead(srange) * 5) / 1024);
     return distance;
 }
